// src/generated/prisma/models/Interest.ts
export type InterestWhereInput = {
  AND?: Prisma.InterestWhereInput | Prisma.InterestWhereInput[]
  OR?: Prisma.InterestWhereInput[]
  NOT?: Prisma.InterestWhereInput | Prisma.InterestWhereInput[]
  id?: Prisma.StringFilter<"Interest"> | string
  amount?: Prisma.IntFilter<"Interest"> | number
  date?: Prisma.DateTimeFilter<"Interest"> | Date | string
  remarks?: Prisma.StringNullableFilter<"Interest"> | string | null
  jinisId?: Prisma.StringNullableFilter<"Interest"> | string | null
  jinisCharaId?: Prisma.StringNullableFilter<"Interest"> | string | null
  personName?: Prisma.StringNullableFilter<"Interest"> | string | null
  createdById?: Prisma.StringFilter<"Interest"> | string
  createdAt?: Prisma.DateTimeFilter<"Interest"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Interest"> | Date | string
  jinis?: Prisma.XOR<Prisma.JinisNullableScalarRelationFilter, Prisma.JinisWhereInput> | null
  jinisChara?: Prisma.XOR<Prisma.JinisCharaNullableScalarRelationFilter, Prisma.JinisCharaWhereInput> | null
  createdBy?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
}